Рабочий вариант авторизации

+++ src/main/java/ru/skypro/homework/service/impl/AdServiceImpl.java
@@ -1,40 +1,44 @@
 package ru.skypro.homework.service.impl;
 
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
-import ru.skypro.homework.dto.AdsDto;
-import ru.skypro.homework.dto.CreateAdsDto;
-import ru.skypro.homework.dto.FullAdsDto;
-import ru.skypro.homework.dto.ResponseWrapperAdsDto;
+import ru.skypro.homework.dto.*;
 import ru.skypro.homework.entity.Ad;
 import ru.skypro.homework.entity.User;
 import ru.skypro.homework.exception.AdsNotFoundException;
 import ru.skypro.homework.exception.UserNotFoundException;
 import ru.skypro.homework.repository.AdRepository;
-import ru.skypro.homework.repository.UserRepository;
 import ru.skypro.homework.service.AdService;
 import ru.skypro.homework.service.UserService;
 import ru.skypro.homework.service.mapper.AdMapper;
 
-import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
+@RequiredArgsConstructor
 public class AdServiceImpl implements AdService {
 
+    private final Role role = Role.ADMIN;
     private final AdRepository adRepository;
-    private final UserRepository userRepository;
     private final UserService userService;
     private final AdMapper adMapper;
 
-    public AdServiceImpl(AdRepository adRepository,
-                         UserRepository userRepository,
-                         UserService userService, AdMapper adMapper) {
-        this.adRepository = adRepository;
-        this.userRepository = userRepository;
-        this.userService = userService;
-        this.adMapper = adMapper;
+    /**
+     * Метод проверяет наличие доступа к объявлению по id
+     * @param id
+     * @see ru.skypro.homework.service.impl.AdServiceImpl
+     */
+    @Override
+    public boolean checkAccess(Integer id) {
+        Ad ad = adRepository.findById(id).orElseThrow(AdsNotFoundException::new);
+        Optional<User> user = userService.findAuthUser();
+        User notOptionalUser = user.get();
+        String currentPrincipalName = notOptionalUser.getUsername();
+        return ad.getAuthor().getUsername().equals(currentPrincipalName)
+                || notOptionalUser.getAuthorities().contains(role);
     }
 
     @Override
@@ -60,7 +64,7 @@ public class AdServiceImpl implements AdService {
 
     @Override
     public boolean removeAdDto(Integer id) {
-        if(adRepository.existsById(id)) {
+        if (checkAccess(id) && adRepository.existsById(id)) {
             adRepository.deleteById(id);
             return true;
         }
@@ -70,9 +74,12 @@ public class AdServiceImpl implements AdService {
     @Override
     public AdsDto updateAdDto(Integer id, CreateAdsDto createAdsDto) {
         Ad ad = adRepository.findById(id).orElseThrow(AdsNotFoundException::new);
-        ad.setDescription(createAdsDto.getDescription());
-        ad.setPrice(createAdsDto.getPrice());
-        ad.setTitle(createAdsDto.getTitle());
+        if (checkAccess(id)) {
+            ad.setDescription(createAdsDto.getDescription());
+            ad.setPrice(createAdsDto.getPrice());
+            ad.setTitle(createAdsDto.getTitle());
+            return adMapper.mapAdToAdDto(adRepository.save(ad));
+        }
         return adMapper.mapAdToAdDto(adRepository.save(ad));
     }
